+++ example/tests/test_builder.py
@@ -2,7 +2,7 @@ import pytest
 from django.contrib.auth import get_user_model
 
 from example.models import Organization, Person
-from example.schema import schema_builder
+from example.schema import schema
 from simple_graphql.django import SchemaBuilder
 from simple_graphql.django.schema.exceptions import SchemaAlreadyBuilt
 
@@ -10,25 +10,23 @@ User = get_user_model()
 
 
 def test_schema_config_loading():
-    person_config = schema_builder.registry.get(Person)
+    person_config = schema.builder.registry.get(Person)
     assert person_config.exclude_fields == ["secret"]
 
-    organization_config = schema_builder.registry.get(Organization)
+    organization_config = schema.builder.registry.get(Organization)
     assert organization_config.default_ordering == "name"
 
-    user_config = schema_builder.registry.get(User)
+    user_config = schema.builder.registry.get(User)
     assert user_config.exclude_fields == ["password"]
 
 
 def test_schema_post_build_registering_errors():
     builder = SchemaBuilder()
-    schema = builder.build_schema()
 
     # Registering should be fine before the schema gets evaluated
     builder.register_model(Person)
 
-    # Registering should fail once the schema has been evaluated
-    # Asserting on the schema will evaluate it
-    assert schema
+    # Registering should fail once the schema has been built
+    builder.build_schema()
     with pytest.raises(SchemaAlreadyBuilt):
         builder.register_model(Organization)