+++ simple_graphql/django/__init__.py
@@ -1,7 +1,8 @@
-from .schema import SchemaBuilder
+from .schema import Schema, SchemaBuilder
 from .types import ModelSchemaConfig
 
 __all__ = [
+    "Schema",
     "SchemaBuilder",
     "ModelSchemaConfig",
 ]